Move KWCOCO ETL code to seperate module
